Updates tags to be prefixed with a 'v'.
Currently we are using just the version number as the tag but our
current practices for existing application is to have the tag be a
version number prefixed but a 'v'.

This also updates the test helping method `with_repo` to have the logic
for removing the newly created directed in an `ensure` block to
guarantee that it runs.