Disables BlockComments Cop in spec_helper.
The Ruby style guide prefers to avoid block comments. Since the
spec_helper's block comment was autogenerated I decided it would be fine
to just ignore the BlockComments Cop around the offending area. I felt
this would be preferable to ignoring the file completely since it could
be updated to add helper method and was also preferable to re-writing
the comments to use a comment per line instead of block quoting.