Fixes issue getting test coverage to work.
Code Climate's test coverage gem requires you ends up in the directory
containing the `.git` directory in the file. Prior to this fix we would
end up in the `spec` directory. This updates the `with_repo` spec helper
to move up an additional directory and end up in the project's root.