Updates `Bumper#current_tag` to use `Tag.current`.
Using `Tag.current` will properly sort git tags numerically instead of
alphabetically. This will resolve issues where we were trying to update to the
wrong version because we were not pulling in the proper, highest, current
version.

We were also able to remove the `NullTag` and "manually" create a version of
this using our `Tag` class with `Tag.new(0, 0, 0)`.